Replace :erlang.byte_size with external_size
external_size function will always show equal to or more bytes than byte_size
function, but we don't need to encode it, so we can still work with iodata.

Also replace @en_function usage with :encode! in truncate_item method to ensure
that once we start truncating we have a string not iodata that can't be sliced.