Prune invalid UTF-8 bytes
If a log message or its metadata value is a binary which is not encoded
as valid UTF-8, then it cannot be encoded into JSON.

E.g. when the following code is executed:
```
Jason.encode!(<<97, 255>>)
```
then the following error is raised:
```
(Jason.EncodeError) invalid byte 0xFF in <<97, 255>>
```

This commit fixes this problem by removing the
invalid bytes from the string.

One corner case where this solution does not help is when the passed
metadata value is a struct which implements an `Encoder`. In such case
the struct is serialized to JSON with the implemented `Encoder` which
may not prune the invalid bytes.

EN-5883